+++ src/main/java/com/sparta/internonboarding/auth/filter/CustomAuthorizationFilter.java
@@ -1,7 +1,7 @@
-package com.sparta.internonboarding.filter;
+package com.sparta.internonboarding.auth.filter;
 
-import com.sparta.internonboarding.jwt.JwtUtil;
-import com.sparta.internonboarding.userdetails.UserDetailsServiceImpl;
+import com.sparta.internonboarding.auth.jwt.JwtUtil;
+import com.sparta.internonboarding.auth.userdetails.UserDetailsServiceImpl;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
