+++ src/main/java/com/sparta/internonboarding/auth/userdetails/UserDetailsImpl.java
@@ -1,4 +1,4 @@
-package com.sparta.internonboarding.userdetails;
+package com.sparta.internonboarding.auth.userdetails;
 
 import com.sparta.internonboarding.user.entity.User;
 import lombok.Getter;
