+++ src/main/java/com/sparta/internonboarding/config/WebSecurityConfig.java
@@ -1,9 +1,9 @@
 package com.sparta.internonboarding.config;
 
-import com.sparta.internonboarding.filter.CustomAuthenticationFilter;
-import com.sparta.internonboarding.filter.CustomAuthorizationFilter;
-import com.sparta.internonboarding.jwt.JwtUtil;
-import com.sparta.internonboarding.userdetails.UserDetailsServiceImpl;
+import com.sparta.internonboarding.auth.filter.CustomAuthenticationFilter;
+import com.sparta.internonboarding.auth.filter.CustomAuthorizationFilter;
+import com.sparta.internonboarding.auth.jwt.JwtUtil;
+import com.sparta.internonboarding.auth.userdetails.UserDetailsServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
